refactor(update): split restart planning helpers
Move update restart plans and platform-specific script builders into a focused helper while preserving simple_update compatibility imports.

Tested: pytest tests/test_simple_update.py tests/test_split_update_timers.py tests/test_secure_temp_update.py tests/test_update_checksum_verification.py; ruff check src tests scripts; pyright.

Confidence: high
